refactored moving tools in to utils.
