+++ cdtea/Visualization/tests/test_coordinates.py
@@ -1,5 +1,5 @@
 from cdtea.generate_flat import generate_flat_2d_space_time
-from cdtea.Visualization.coordinates import toroidal_coordinates, nearest
+from cdtea.Visualization.coordinates import toroidal_coordinates
 import numpy as np
 
 
@@ -16,18 +16,4 @@ class TestSpatialOrdering:
             assert type(coords[c][0]) is float
             assert type(coords[c][1]) is float
 
-    def test_nearest(self):
-        """
-            for each layer test that it is the correct size
-            and that all nodes are contained within two edges.
 
-        """
-        ref = np.array([.8, .5]) * 2 * np.pi
-        pnt = np.array([.1, .6]) * 2 * np.pi
-        result = np.array([1.1, .6]) * 2 * np.pi
-        assert np.allclose(nearest(ref, pnt), result)
-
-        ref = np.array([.2, .1]) * 2 * np.pi
-        pnt = np.array([.9, .8]) * 2 * np.pi
-        result = np.array([-.1, -.2]) * 2 * np.pi
-        assert np.allclose(nearest(ref, pnt), result)
