+++ cdtea/tests/test_generate_flat.py
@@ -2,7 +2,7 @@ from cdtea import simplicial
 from cdtea import generate_flat
 from collections import defaultdict
 from cdtea.tests.valid_triangulation import is_valid
-from cdtea.util.TimeIndex import time_sep
+from cdtea.util.triangulation_utils import time_sep
 
 
 class TestGenerateFlat:
